Добавлена тестовая фильтрация для таблицы

+++ app/routes/post.py
@@ -1,8 +1,8 @@
-from flask import Blueprint, render_template, request, redirect
+from flask import Blueprint, render_template, request, redirect, abort
 from flask_login import login_required, current_user
 
 from ..models.user import User
-from ..forms import StudentForm
+from ..forms import StudentForm, TeacherForm
 from ..extensions import db
 from ..models.post import Post
 
@@ -10,8 +10,16 @@ post = Blueprint('post', __name__)
 
 @post.route('/', methods=['POST', 'GET'])
 def all():
-    posts = Post.query.order_by(Post.date.desc()).all()
-    return render_template('post/all.html', posts=posts, user=User)
+    form = TeacherForm()
+    form.teacher.choices = [t.name for t in User.query.filter_by(status='teacher')]
+
+    if request.method == "POST":
+        teacher = request.form.get('teacher')
+        teacher_id = User.query.filter_by(name=teacher).first().id
+        posts = Post.query.filter_by(teacher=teacher_id).order_by(Post.date.desc()).all()
+    else:
+        posts = Post.query.order_by(Post.date.desc()).limit(20).all()
+    return render_template('post/all.html', posts=posts, user=User, form=form)
 
 @post.route('/post/create', methods=['POST', 'GET'])
 @login_required
@@ -39,27 +47,39 @@ def create():
 @login_required
 def update(id):
     post = Post.query.get(id)
-    if request.method == 'POST':
-        post.teacher = request.form.get('teacher')
-        post.subject = request.form.get('subject')
-        post.student = request.form.get('student')
 
-        try:
-            db.session.commit()
-            return redirect('/')
-        except Exception as e:
-            print(str(e))        
+    if post.author.id == current_user.id:
+        form = StudentForm()
+        form.student.data = User.query.filter_by(id=post.student).first().name
+        form.student.choices = [s.name for s in User.query.filter_by(status='user')]
+        if request.method == 'POST':
+            post.subject = request.form.get('subject')
+            student = request.form.get('student')
+
+            post.student = User.query.filter_by(name=student).first().id
+
+            try:
+                db.session.commit()
+                return redirect('/')
+            except Exception as e:
+                print(str(e))        
+        else:
+            return render_template('post/update.html', post=post, form=form)
     else:
-        return render_template('post/update.html', post=post)
+        abort(403)
 
 @post.route('/post/<int:id>/delete', methods=['POST', 'GET'])
 @login_required
 def delete(id):
     post = Post.query.get(id)
-    try:
-        db.session.delete(post)
-        db.session.commit()
-        return redirect('/')
-    except Exception as e:
-        print(str(e))
-        return str(e)        
+
+    if post.author.id == current_user.id:
+        try:
+            db.session.delete(post)
+            db.session.commit()
+            return redirect('/')
+        except Exception as e:
+            print(str(e))
+            return str(e)        
+    else:
+        abort(403)
